Eliminate the numerical problems when the number of microphones increases.
When the number of microphones increases(M=6, for example), the coherence matrix DC(:,:,k) tends to be a non-positive matrix due to the numerical problem(It will have some very small negative eigenvalues especially in the low-frequency range) and it will influence the following code. We may add a small positive number to its diagonal elements to eliminate this numerical problem.